@@ -1,40 +1,40 @@
 module github.com/go-acme/lego/v4
 
-go 1.17
+go 1.18
 
 // github.com/exoscale/egoscale v1.19.0 => It is an error, please don't use it.
-// github.com/linode/linodego v1.0.0 => It is an error, please don't use it.
 require (
 	cloud.google.com/go v0.54.0
 	github.com/Azure/azure-sdk-for-go v32.4.0+incompatible
-	github.com/Azure/go-autorest/autorest v0.11.19
-	github.com/Azure/go-autorest/autorest/azure/auth v0.5.8
+	github.com/Azure/go-autorest/autorest v0.11.24
+	github.com/Azure/go-autorest/autorest/azure/auth v0.5.11
 	github.com/Azure/go-autorest/autorest/to v0.4.0
 	github.com/Azure/go-autorest/autorest/validation v0.3.1 // indirect
-	github.com/BurntSushi/toml v0.3.1
+	github.com/BurntSushi/toml v1.2.0
 	github.com/OpenDNS/vegadns2client v0.0.0-20180418235048-a3fa4a771d87
-	github.com/akamai/AkamaiOPEN-edgegrid-golang v1.1.1
-	github.com/aliyun/alibaba-cloud-sdk-go v1.61.1183
+	github.com/akamai/AkamaiOPEN-edgegrid-golang v1.2.1
+	github.com/aliyun/alibaba-cloud-sdk-go v1.61.1755
 	github.com/aws/aws-sdk-go v1.39.0
-	github.com/cenkalti/backoff/v4 v4.1.1
-	github.com/cloudflare/cloudflare-go v0.20.0
+	github.com/cenkalti/backoff/v4 v4.1.3
+	github.com/civo/civogo v0.3.11
+	github.com/cloudflare/cloudflare-go v0.49.0
 	github.com/cpu/goacmedns v0.1.1
-	github.com/dnsimple/dnsimple-go v0.70.1
-	github.com/exoscale/egoscale v0.67.0
+	github.com/dnsimple/dnsimple-go v0.71.1
+	github.com/exoscale/egoscale v0.90.0
 	github.com/google/go-querystring v1.1.0
-	github.com/gophercloud/gophercloud v0.16.0
+	github.com/gophercloud/gophercloud v1.0.0
 	github.com/gophercloud/utils v0.0.0-20210216074907-f6de111f2eae
 	github.com/iij/doapi v0.0.0-20190504054126-0bbf12d6d7df
 	github.com/infobloxopen/infoblox-go-client v1.1.1
 	github.com/labbsr0x/bindman-dns-webhook v1.0.2
-	github.com/linode/linodego v0.31.1
+	github.com/linode/linodego v1.9.1
 	github.com/liquidweb/liquidweb-go v1.6.3
-	github.com/mattn/go-isatty v0.0.14
-	github.com/miekg/dns v1.1.47
+	github.com/mattn/go-isatty v0.0.16
+	github.com/miekg/dns v1.1.50
 	github.com/mimuret/golang-iij-dpf v0.7.1
-	github.com/mitchellh/mapstructure v1.4.1
+	github.com/mitchellh/mapstructure v1.5.0
 	github.com/namedotcom/go v0.0.0-20180403034216-08470befbe04
-	github.com/nrdcg/auroradns v1.0.1
+	github.com/nrdcg/auroradns v1.1.0
 	github.com/nrdcg/desec v0.6.0
 	github.com/nrdcg/dnspod-go v0.4.0
 	github.com/nrdcg/freemyip v0.2.0
