@@ -6,6 +6,7 @@ import (
 	"crypto/rand"
 	"crypto/rsa"
 	"encoding/pem"
+	"regexp"
 	"testing"
 	"time"
 
@@ -112,7 +113,9 @@ func TestPEMEncode(t *testing.T) {
 
 	data := PEMEncode(key)
 	require.NotNil(t, data)
-	assert.Len(t, data, 127)
+
+	exp := regexp.MustCompile(`^-----BEGIN RSA PRIVATE KEY-----\s+\S{60,}\s+-----END RSA PRIVATE KEY-----\s+`)
+	assert.Regexp(t, exp, string(data))
 }
 
 func TestParsePEMCertificate(t *testing.T) {
