@@ -283,7 +283,7 @@ func (c *Client) doRequest(method string, uri *url.URL) (json.RawMessage, error)
 		return nil, errors.New(toUnreadableBodyMessage(req, content))
 	}
 
-	if resp.StatusCode != 200 {
+	if resp.StatusCode != http.StatusOK {
 		return nil, fmt.Errorf("invalid code (%d), error: %s", resp.StatusCode, content)
 	}
 
@@ -317,19 +317,20 @@ func toUnreadableBodyMessage(req *http.Request, rawBody []byte) string {
 
 // Rounds the given TTL in seconds to the next accepted value.
 // Accepted TTL values are:
-//  - 60 = 1 minute
-//  - 300 = 5 minutes
-//  - 900 = 15 minutes
-//  - 1800 = 30 minutes
-//  - 3600 = 1 hour
-//  - 21600 = 6 hours
-//  - 43200 = 12 hours
-//  - 86400 = 1 day
-//  - 172800 = 2 days
-//  - 259200 = 3 days
-//  - 604800 = 1 week
-//  - 1209600 = 2 weeks
-//  - 2592000 = 1 month
+//   - 60 = 1 minute
+//   - 300 = 5 minutes
+//   - 900 = 15 minutes
+//   - 1800 = 30 minutes
+//   - 3600 = 1 hour
+//   - 21600 = 6 hours
+//   - 43200 = 12 hours
+//   - 86400 = 1 day
+//   - 172800 = 2 days
+//   - 259200 = 3 days
+//   - 604800 = 1 week
+//   - 1209600 = 2 weeks
+//   - 2592000 = 1 month
+//
 // See https://www.cloudns.net/wiki/article/58/ for details.
 func ttlRounder(ttl int) int {
 	for _, validTTL := range []int{60, 300, 900, 1800, 3600, 21600, 43200, 86400, 172800, 259200, 604800, 1209600} {
