@@ -168,7 +168,7 @@ func TestDNSProvider_FindZoneAndRecordName(t *testing.T) {
 		t.Run(test.desc, func(t *testing.T) {
 			t.Parallel()
 
-			zone, recordName, err := provider.FindZoneAndRecordName(test.fqdn, test.domain)
+			zone, recordName, err := provider.findZoneAndRecordName(test.fqdn, test.domain)
 			require.NoError(t, err)
 			assert.Equal(t, test.expected.zone, zone)
 			assert.Equal(t, test.expected.recordName, recordName)
